Try-Catch agregados al helper fetch - Listo

+++ frontend/src/actions/loginActions.ts
@@ -9,6 +9,7 @@ import {
 } from './loginActionsTypes';
 import Swal from 'sweetalert2';
 import { fetchConToken, fetchSinToken } from '../helpers/fetch';
+import { Toast } from '../helpers/swalAlert';
 
 //Login
 export const startLogin = (email: string, password: string) => {
@@ -17,32 +18,42 @@ export const startLogin = (email: string, password: string) => {
 		//dispatch para cambiar loading a true
 		dispatch({ type: AUTH_START_LOADING });
 		//Peticion Fetch a la API para hacer login
-		const respuesta = await fetchSinToken(
-			'users/authenticate',
-			{ username: email, password: password },
-			'POST'
-		);
-		//.json() a la respuesta
-		const body = await respuesta.json();
 
-		//Condicion si existe un id
-		if (body.ok) {
-			//Se guarda el token en localStorage como gersa-tkn
-			localStorage.setItem('gersa-tkn', body.data.token);
-			const time = new Date().getTime();
-			//Se guarda el tiempo en el que se guardo el token en localStorage como gersa-tkn-init-date
-			localStorage.setItem('gersa-tkn-init-date', time.toString());
-			localStorage.setItem('gersaUserName', body.data.username);
-			//Se asugna el cuerpo de la respuesta a usuario
-			const usuario: Usuario = body.data;
-			//dispatch que guarda al usuario obtenido en el reducer
-			dispatch({
-				type: AUTH_SUCCESS,
-				payload: { usuario },
+		try {
+			const respuesta = await fetchSinToken(
+				'users/authenticate',
+				{ username: email, password: password },
+				'POST'
+			);
+			//.json() a la respuesta
+			const body = await respuesta?.json();
+
+			//Condicion si existe un id
+			if (body.ok) {
+				//Se guarda el token en localStorage como gersa-tkn
+				localStorage.setItem('gersa-tkn', body.data.token);
+				const time = new Date().getTime();
+				//Se guarda el tiempo en el que se guardo el token en localStorage como gersa-tkn-init-date
+				localStorage.setItem('gersa-tkn-init-date', time.toString());
+				localStorage.setItem('gersaUserName', body.data.username);
+				//Se asugna el cuerpo de la respuesta a usuario
+				const usuario: Usuario = body.data;
+				//dispatch que guarda al usuario obtenido en el reducer
+				dispatch({
+					type: AUTH_SUCCESS,
+					payload: { usuario },
+				});
+			} else {
+				//Mensaje de error proveniente de la API
+				Swal.fire('Error', body.message, 'error');
+				dispatch({ type: AUTH_LOADING_FINISH });
+			}
+		} catch (error) {
+			console.log(error);
+			Toast.fire({
+				icon: 'error',
+				title: '¡Ups! Algo salió mal. <br/> Por favor, Intenta de nuevo!',
 			});
-		} else {
-			//Mensaje de error proveniente de la API
-			Swal.fire('Error', body.message, 'error');
 			dispatch({ type: AUTH_LOADING_FINISH });
 		}
 	};
@@ -54,31 +65,33 @@ export const startChecking = () => {
 	return async (dispatch: Dispatch<AuthDispatchTypes>) => {
 		//Se recupera el token guardado el localStorage
 		const respuesta = await fetchConToken('users/renew', {}, 'GET');
-		const body = await respuesta.json();
+		const body = await respuesta?.json();
 
-		//Condicion si existe un id
-		if (body.ok) {
-			//Se guarda el token en localStorage como gersa-tkn
-			localStorage.setItem('gersa-tkn', body.data.token);
-			const time = new Date().getTime();
-			//Se guarda el tiempo en el que se guardo el token en localStorage como gersa-tkn-init-date
-			localStorage.setItem('gersa-tkn-init-date', time.toString());
-			localStorage.setItem('gersaUserName', body.data.username);
-			//Se asugna el cuerpo de la respuesta a usuario
-			const usuario: Usuario = body.data;
-			//dispatch que guarda al usuario obtenido en el reducer
-			dispatch({
-				type: AUTH_SUCCESS,
-				payload: { usuario },
-			});
-		} else {
-			//Mensaje de error proveniente de la API
-			console.log(body.message);
-			dispatch({ type: AUTH_LOADING_FINISH });
-			//Limpia localStorage para cerrar la sesion si el toquen no fue valido
-			localStorage.clear();
-			dispatch({ type: AUTH_LOGOUT });
-		}
+		if (body) {
+			//Condicion si existe un id
+			if (body.ok) {
+				//Se guarda el token en localStorage como gersa-tkn
+				localStorage.setItem('gersa-tkn', body.data.token);
+				const time = new Date().getTime();
+				//Se guarda el tiempo en el que se guardo el token en localStorage como gersa-tkn-init-date
+				localStorage.setItem('gersa-tkn-init-date', time.toString());
+				localStorage.setItem('gersaUserName', body.data.username);
+				//Se asugna el cuerpo de la respuesta a usuario
+				const usuario: Usuario = body.data;
+				//dispatch que guarda al usuario obtenido en el reducer
+				dispatch({
+					type: AUTH_SUCCESS,
+					payload: { usuario },
+				});
+			} else {
+				//Mensaje de error proveniente de la API
+				console.log(body.message);
+				dispatch({ type: AUTH_LOADING_FINISH });
+				//Limpia localStorage para cerrar la sesion si el toquen no fue valido
+				localStorage.clear();
+				dispatch({ type: AUTH_LOGOUT });
+			}
+		} else return;
 	};
 };
